Modularize functions into AudioLibrary Class. Implemented audio downloading function.
